requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
codeclimate-test-reporter==0.2.3
colorama==0.4.0
coverage==4.5.1
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-JWT-Extended==3.13.1
Flask-RESTful==0.3.6
flask-restplus==0.12.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
nose==1.3.7
passlib==1.7.1
pluggy==0.7.1
py==1.7.0
PyJWT==1.4.2
pylint==2.1.1
pytest==3.8.2
pytest-cov==2.6.0
pytz==2018.5
requests==2.20.0
six==1.11.0
SQLAlchemy==1.2.12
typed-ast==1.1.0
urllib3==1.24
Werkzeug==0.14.1
wrapt==1.10.11
